@@ -19,15 +19,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.0",
-    "@angular/common": "^6.0.0",
-    "@angular/compiler": "^6.0.0",
-    "@angular/core": "^6.0.0",
-    "@angular/forms": "^6.0.0",
-    "@angular/http": "^6.0.0",
-    "@angular/platform-browser": "^6.0.0",
-    "@angular/platform-browser-dynamic": "^6.0.0",
-    "@angular/router": "^6.0.0",
+    "@angular/animations": "^6.1.2",
+    "@angular/common": "^6.1.2",
+    "@angular/compiler": "^6.1.2",
+    "@angular/core": "^6.1.2",
+    "@angular/forms": "^6.1.2",
+    "@angular/http": "^6.1.2",
+    "@angular/platform-browser": "^6.1.2",
+    "@angular/platform-browser-dynamic": "^6.1.2",
+    "@angular/router": "^6.1.2",
     "@fortawesome/fontawesome": "^1.1.8",
     "@fortawesome/fontawesome-free-regular": "^5.0.13",
     "@fortawesome/fontawesome-free-solid": "^5.0.6",
@@ -48,9 +48,9 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.7.0-beta.1",
-    "@angular/cli": "~6.0.0",
-    "@angular/compiler-cli": "^6.0.0",
-    "@angular/language-service": "^6.0.0",
+    "@angular/cli": "~6.1.2",
+    "@angular/compiler-cli": "^6.1.2",
+    "@angular/language-service": "^6.1.2",
     "@compodoc/compodoc": "^1.1.3",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
