@@ -1,12 +1,14 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable, Subject, BehaviorSubject, ReplaySubject, throwError } from 'rxjs';
-import { tap, catchError } from 'rxjs/operators';
+import { tap, catchError, switchMap, map } from 'rxjs/operators';
 import { AccessTokenResponse } from '~features/auth/interfaces/access-token-response.interface';
 import { RefreshTokenResponse } from '~features/auth/interfaces/refresh-token-response.interface';
 import { LocalStorageService } from '~shared/local-storage';
 import { AccessTokenState } from '~features/auth/interfaces';
 import { AuthModule } from '~features/auth/auth.module';
+import { environment } from 'environments/environment.prod';
+import { GuestAccessTokenState } from '~features/auth/interfaces/guest-access-token-state.interface';
 
 const ACCESS_TOKEN_NAME = 'accessToken';
 const REFRESH_TOKEN_NAME = 'refreshToken';
@@ -18,6 +20,8 @@ const REFRESH_TOKEN_NAME = 'refreshToken';
 export class TokenService {
 	private _accessToken$ = new ReplaySubject<AccessTokenState>(1);
 	accessToken$ = this._accessToken$.asObservable();
+	private _guestAccessToken$ = new ReplaySubject<GuestAccessTokenState>(1);
+	guestAccessToken$ = this._guestAccessToken$.asObservable();
 	// having the token accessible synchronously makes things easier in other places
 	accessTokenSync: AccessTokenState;
 	// timeout variable. The timeout will refresh the access token.
@@ -31,7 +35,7 @@ export class TokenService {
 	clearTokens(): void {
 		this.localStorageSrv.remove(REFRESH_TOKEN_NAME);
 		this.localStorageSrv.remove(ACCESS_TOKEN_NAME);
-		this._accessToken$.next({ token: null, token_data: null });
+		this._accessToken$.next({ invalidated: true });
 		this.stopTimer();
 	}
 
@@ -50,39 +54,57 @@ export class TokenService {
 		this.refreshToken = this.localStorageSrv.getItem(REFRESH_TOKEN_NAME);
 
 		if (this.refreshToken)
-			this.fetchAccessToken().subscribe();
+			this.generateAccessToken(this.refreshToken).subscribe();
 		else
-			this._accessToken$.next({ token: null, token_data: null });
+			this._accessToken$.next({ invalidated: true });
 
 	}
 
-	generateAccessToken(refreshToken: RefreshTokenResponse) {
+	getGuestAccessToken(token: string): Observable<GuestAccessTokenState> {
+		return this.http.get<RefreshTokenResponse>(`https://ros-dev3.showsourcing.com/token/${token}`).pipe(
+			switchMap((refreshToken: any) => this.fetchAccessToken(refreshToken).pipe(
+				map(accessTokenResponse => ({
+					token: accessTokenResponse.user_token.token,
+					token_data: accessTokenResponse.user_token.token_data,
+					realm: refreshToken.realm as any,
+				}))
+			)),
+			tap(guestToken => {
+				this._guestAccessToken$.next(guestToken);
+			})
+		);
+	}
+
+	revokeGuestAccessToken() {
+		this._guestAccessToken$.next({ invalidated: true });
+	}
+
+	generateAccessToken(refreshToken: RefreshTokenResponse): Observable<AccessTokenResponse> {
 		this.localStorageSrv.setItem(REFRESH_TOKEN_NAME, refreshToken);
 		this.refreshToken = refreshToken;
-		return this.fetchAccessToken();
-	}
-
-	private fetchAccessToken(): Observable<AccessTokenResponse> {
-		const accessObj = {
-			app_id: '',
-			provider: 'realm',
-			data: this.refreshToken.refresh_token.token,
-		};
-		return this.http.post<AccessTokenResponse>('api/auth', accessObj).pipe(
-			tap(token => this.onNewAccessToken(token)),
+		return this.fetchAccessToken(refreshToken).pipe(
+			tap(accessToken => this.onNewAccessToken(accessToken)),
 			catchError(e => {
-				this._accessToken$.next({ token: null, token_data: null });
+				this._accessToken$.next({ invalidated: true });
 				return throwError(e);
 			})
 		);
 	}
 
+	fetchAccessToken(refreshToken): Observable<AccessTokenResponse> {
+		const accessObj = {
+			app_id: '',
+			provider: 'realm',
+			data: refreshToken.refresh_token.token,
+		};
+		return this.http.post<AccessTokenResponse>(`${environment.apiUrl}/auth`, accessObj);
+	}
+
 	/** when a new access token arrives */
 	private onNewAccessToken(accessToken: AccessTokenResponse) {
 		const accessTokenState = {
-			pending: false,
 			token: accessToken.user_token.token,
-			token_data: accessToken.user_token.token_data
+			token_data: accessToken.user_token.token_data,
 		};
 		this._accessToken$.next(accessTokenState);
 		this.localStorageSrv.setItem(ACCESS_TOKEN_NAME, accessTokenState);
@@ -95,7 +117,7 @@ export class TokenService {
 		const errorDelta = 5 * 60 * 1000;
 		const delta = (1000 * accessTokenState.token_data.expires) - Date.now();
 		this.timer = setTimeout(_ => {
-			this.fetchAccessToken();
+			this.fetchAccessToken(this.refreshToken);
 		}, (delta - errorDelta));
 	}
 
