@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
 import { Apollo } from 'apollo-angular';
 import { HttpLink } from 'apollo-angular-link-http';
-import { distinctUntilChanged, filter, map } from 'rxjs/operators';
+import { distinctUntilChanged, filter, map, switchMap } from 'rxjs/operators';
 import { AuthenticationService } from '~features/auth/services/authentication.service';
 import { TokenService } from '~features/auth/services/token.service';
 import { Team } from '~models/team.model';
-import { ApolloStateService } from '~shared/apollo/services/apollo-state.service';
+import { ApolloStateService } from './apollo-state.service';
 import { AbstractApolloInitializer } from '~shared/apollo/services/initializers/abstract-apollo-initializer.class';
 import { log } from '~utils/log';
 import { TeamPickerService } from '~features/pick-a-team/services/team-picker.service';
@@ -16,39 +16,37 @@ export class TeamClientInitializer extends AbstractApolloInitializer {
 
 	constructor(
 		protected apollo: Apollo,
-		protected tokenSrv: TokenService,
 		protected link: HttpLink,
+		protected tokenSrv: TokenService,
 		protected apolloState: ApolloStateService,
 		protected authSrv: AuthenticationService,
 		protected teamPicker: TeamPickerService
 	) {
-		super(apollo, tokenSrv, link, authSrv, true);
+		super(apollo, link);
 	}
 
 	init() {
-		// when the the user has selected a team we initialize the team client
+		// when the user has selected a team we initialize the team client
 		this.teamPicker.selectedTeam$
 			.pipe(
 				filter(t => !!t),
 				distinctUntilChanged((x, y) => x.id === y.id),
-		).subscribe(team => this.initTeamClient(team));
+				switchMap(team => this.getRealmUri(team.realmServerName, team.realmPath))
+			).subscribe(uri => this.initTeamClient(uri, this.tokenSrv.accessTokenSync.token));
 
-		// when authenticated we start user client
+		// when authenticated we start team client
 		this.authSrv.authState$.pipe(
 			map(authState => authState.authenticated),
-		).subscribe(authenticated => {
-			if (!authenticated)
-				this.resetClient();
-		});
+			filter(authenticated => !authenticated)
+		).subscribe(authenticated => this.resetClient());
+
 	}
 
 
 	/** initialize apollo team client */
-	private async initTeamClient(team: Team) {
+	private async initTeamClient(uri: string, token) {
 		try {
-			const realm = await this.getRealm(team.realmServerName);
-			const uris = this.getUris(realm.httpsPort, realm.hostname, team.realmPath);
-			this.createClient(uris.httpUri, uris.wsUri);
+			this.createClient(uri, undefined, token);
 			this.apolloState.setTeamClientReady();
 		} catch (e) {
 			log.error(e);
