@@ -1,51 +1,55 @@
 import { Injectable } from '@angular/core';
 import { Apollo } from 'apollo-angular';
 import { HttpLink } from 'apollo-angular-link-http';
-import { distinctUntilChanged, filter, first, switchMap } from 'rxjs/operators';
+import { distinctUntilChanged, first, switchMap } from 'rxjs/operators';
 import { AuthenticationService } from '~features/auth/services/authentication.service';
 import { TokenService } from '~features/auth/services/token.service';
 import { User } from '~models/user.model';
-import { ALL_USER_CLIENT, USER_CLIENT } from '~shared/apollo/services/apollo-endpoints.const';
-import { ApolloStateService } from '~shared/apollo/services/apollo-state.service';
+import { ALL_USER_CLIENT, USER_CLIENT } from './client-names.const';
+import { ApolloStateService } from './apollo-state.service';
 import { ApolloWrapper } from '~shared/apollo/services/apollo-wrapper.service';
 import { AbstractApolloInitializer } from '~shared/apollo/services/initializers/abstract-apollo-initializer.class';
 import { ClientInitializerQueries } from '~shared/apollo/services/initializers/initializer-queries';
 import { log } from '~utils/log';
-import { from, of } from 'rxjs';
+import { filter } from 'rxjs/operators';
+import { Observable } from 'rxjs';
 
 @Injectable({ providedIn: 'root' })
 export class UserClientInitializer extends AbstractApolloInitializer {
 
 	constructor(
 		protected apollo: Apollo,
-		protected tokenSrv: TokenService,
 		protected link: HttpLink,
+		protected tokenSrv: TokenService,
 		protected apolloState: ApolloStateService,
 		protected authSrv: AuthenticationService,
 		private wrapper: ApolloWrapper
 	) {
-		super(apollo, tokenSrv, link, authSrv, true);
+		super(apollo, link);
 	}
 
 	init() {
 		// when authenticated we start user client
-		return this.authSrv.authState$.pipe(
-			distinctUntilChanged((x, y) => x.userId === y.userId)
-		).subscribe(authState => {
-			if (authState.authenticated)
-				this.initUserClient(authState.userId);
-			else
-				this.resetClient();
-		});
+		this.authSrv.authState$.pipe(
+			distinctUntilChanged((x, y) => x.userId === y.userId),
+			filter(authState => authState.authenticated),
+			switchMap(authState => this.getUser(authState.userId)),
+			switchMap(user => super.getRealmUri(user.realmServerName, user.realmPath))
+		).subscribe(uri => this.initUserClient(uri, this.tokenSrv.accessTokenSync.token));
+
+
+		// when unauthenticated we reset user client
+		this.authSrv.authState$.pipe(
+			distinctUntilChanged((x, y) => x.userId === y.userId),
+			filter(authState => !authState.authenticated),
+		).subscribe(user => this.resetClient());
+
 	}
 
 	/** create the user client  */
-	private async initUserClient(id: string) {
+	private initUserClient(uri: string, token: string) {
 		try {
-			const user = await this.getUser(id);
-			const realm = await super.getRealm(user.realmServerName);
-			const userUris = super.getUris(realm.httpsPort, realm.hostname, user.realmPath);
-			super.createClient(userUris.httpUri, userUris.wsUri, USER_CLIENT);
+			super.createClient(uri, USER_CLIENT, token);
 			this.apolloState.setUserClientReady();
 		} catch (e) {
 			log.error(e);
@@ -54,14 +58,14 @@ export class UserClientInitializer extends AbstractApolloInitializer {
 	}
 
 	/** gets user from all-users realm */
-	private async getUser(id: string): Promise<User> {
+	private getUser(id: string): Observable<User> {
 		// we use a query here because we need to get the user once from all_user client
 		return this.wrapper.use(ALL_USER_CLIENT).selectOne({
 			gql: ClientInitializerQueries.selectUser,
 			id
 		}).pipe(
 			first()
-		).toPromise();
+		);
 	}
 
 	private resetClient() {
