@@ -30,11 +30,11 @@ export class AuthHttpService {
 	}
 
 	register(credentials: { email: string; password: string, firstName: string, lastName: string }): Observable<any> {
-		return this.http.post(`signup/user`, credentials);
+		return this.http.post(`${environment.apiUrl}/signup/user`, credentials);
 	}
 
 	resetPw(email: string): Observable<any> {
-		return this.http.post(`/api/password/${email}/reset`, {});
+		return this.http.post(`${environment.apiUrl}/api/password/${email}/reset`, {});
 	}
 
 }
