@@ -3,8 +3,8 @@ import { combineLatest, Observable, ReplaySubject, Subject } from 'rxjs';
 import { distinctUntilChanged, filter, map, shareReplay, switchMap, skip } from 'rxjs/operators';
 import { AuthenticationService } from '~features/auth/services/authentication.service';
 import { Team } from '~models';
-import { USER_CLIENT } from '~shared/apollo/services/apollo-endpoints.const';
-import { ApolloStateService } from '~shared/apollo/services/apollo-state.service';
+import { USER_CLIENT } from '~shared/apollo/services/initializers/client-names.const';
+import { ApolloStateService } from '~shared/apollo/services/initializers/apollo-state.service';
 import { ApolloWrapper } from '~shared/apollo/services/apollo-wrapper.service';
 import { LocalStorageService } from '~shared/local-storage';
 import { TeamQueries } from '~global-services/team/team.queries';
