@@ -7,7 +7,7 @@ import { TeamService } from '~global-services';
 import { SelectParams } from '~global-services/_global/select-params';
 import { Sort } from '~shared/table/components/sort.interface';
 import { ApolloWrapper } from '~shared/apollo/services/apollo-wrapper.service';
-import { ApolloStateService } from '~shared/apollo/services/apollo-state.service';
+import { ApolloStateService } from '~shared/apollo/services/initializers/apollo-state.service';
 import { LocalStorageService } from '~shared/local-storage';
 import { TeamPickerService } from '~features/pick-a-team/services/team-picker.service';
 
