@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { CanActivate, CanActivateChild } from '@angular/router';
 import { Observable } from 'rxjs';
 import { filter, map, tap } from 'rxjs/operators';
-import { ApolloStateService } from '~shared/apollo/services/apollo-state.service';
+import { ApolloStateService } from '~shared/apollo/services/initializers/apollo-state.service';
 import { log, LogColor } from '~utils';
 
 @Injectable({
@@ -15,7 +15,6 @@ export class GloabalClientsReadyGuardService implements CanActivate, CanActivate
 	canActivate(): boolean | Observable<boolean> | Promise<boolean> {
 		return this.apolloState.globalClientsReady$.pipe(
 			tap(d => log.debug('%c GlobalClientsReadyGuard', LogColor.GUARD, d.ready)),
-			filter(state => !state.pending),
 			map(state => state.ready)
 		);
 	}
