@@ -3,12 +3,13 @@ import { Apollo } from 'apollo-angular';
 import { HttpLink } from 'apollo-angular-link-http';
 import { InMemoryCache } from 'apollo-cache-inmemory';
 import { from, split, ApolloLink } from 'apollo-link';
+import { Observable } from 'rxjs';
 import { WebSocketLink } from 'apollo-link-ws';
 import { getMainDefinition } from 'apollo-utilities';
 import { first, map } from 'rxjs/operators';
 import { TokenService } from '~features/auth';
 import { AuthenticationService } from '~features/auth/services/authentication.service';
-import { GLOBAL_CONSTANT_CLIENT } from '~shared/apollo/services/apollo-endpoints.const';
+import { GLOBAL_CONSTANT_CLIENT } from './client-names.const';
 import { cleanTypenameLink } from '~shared/apollo/services/clean.typename.link';
 import { ClientInitializerQueries } from '~shared/apollo/services/initializers/initializer-queries';
 import { environment } from 'environments/environment.prod';
@@ -18,48 +19,31 @@ export abstract class AbstractApolloInitializer {
 
 	constructor(
 		protected apollo: Apollo,
-		protected tokenSrv: TokenService,
 		protected httpLink: HttpLink,
-		protected authSrv: AuthenticationService,
-		protected clearOnLogout: boolean
 	) { }
 
 	/**
  	* to create the uri we need to concatena every parts we got from different DB's.
 	*/
-	protected getUris(port: number | string, hostName: string, path: string): { httpUri: string, wsUri: string } {
-		const httpUri = `https://${hostName}:${port}/graphQL/${encodeURIComponent(path)}`;
-		// uri for websocket
-		const wsUri = `wss://${hostName}:${port}/graphQL/${encodeURIComponent(path)}`;
-		return { httpUri: httpUri.toString(), wsUri: wsUri.toString() };
+	protected getUri(port: number | string, hostName: string, path: string): string {
+		return `wss://${hostName}:${port}/graphQL/${encodeURIComponent(path)}`;
 	}
 
 	/**
 	 * gets a realm given a realm name
 	 */
-	protected async getRealm(realmName: string): Promise<{ hostname: string, httpsPort: string }> {
+	protected getRealmUri(realmName: string, path?: string): Observable<string> {
 		return this.apollo.use(GLOBAL_CONSTANT_CLIENT).subscribe({
 			query: ClientInitializerQueries.selectRealmHostName,
 			variables: { query: `name == "${realmName}"` }
 		}).pipe(
 			first(),
-			map((r: any) => r.data.realmServers[0])
-		).toPromise();
+			map((r: any) => r.data.realmServers[0]),
+			map(r => this.getUri(r.httpsPort, r.hostname, path))
+		);
 	}
 
-	protected createClient(httpUri: string, wsUri: string, name?: string) {
-		// Create an http link:
-		let token;
-		let headers;
-		if (this.tokenSrv.accessTokenSync) {
-			token = this.tokenSrv.accessTokenSync.token;
-			headers = new HttpHeaders({ Authorization: token });
-		}
-
-		const http = this.httpLink.create({
-			uri: httpUri,
-			headers
-		});
+	protected createClient(wsUri: string, name?: string, token?: string) {
 
 		// Create a WebSocket link:
 		const connectionParams = (token ? { token } : undefined);
