@@ -1,4 +1,9 @@
 import { Component, OnInit, ChangeDetectionStrategy, ViewChild } from '@angular/core';
+import { AutoUnsub } from '~utils';
+import { ActivatedRoute } from '@angular/router';
+import { takeUntil } from 'rxjs/operators';
+import { GuestClientInitializer } from '~shared/apollo';
+import { TokenService } from '~features/auth';
 
 @Component({
 	selector: 'rfq-template-app',
@@ -6,16 +11,13 @@ import { Component, OnInit, ChangeDetectionStrategy, ViewChild } from '@angular/
 	styleUrls: ['./rfq-template.component.scss'],
 	changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class RfqTemplateComponent implements OnInit {
-	@ViewChild('main') mainRef;
-
-	constructor() { }
+export class RfqTemplateComponent extends AutoUnsub implements OnInit {
+	constructor(
+	) {
+		super();
+	}
 
 	ngOnInit() {
 	}
 
-	onActivate(event) {
-		// this.mainRef.nativeElement.scrollIntoView();
-		window.scrollTo(0, 0);
-	}
 }
