@@ -2,8 +2,8 @@ import { LogLevel } from '~utils/logger/log-level';
 
 export const environment = {
 	production: true,
+	graphqlUrl: 'wss://ros-dev3.showsourcing.com:9443/graphql',
 	apiUrl: 'https://ros-dev3.showsourcing.com:9443',
-	signupUrl: 'https://ros-dev3.showsourcing.com/signup',
 	LOG_LEVEL: LogLevel.ERROR,
 };
 
